requirements.txt
Flask
Flask-SQLAlchemy
flask-cors
openai
python-dotenv